begitulah

+++ dto/business_dto.go
@@ -5,8 +5,8 @@ type CreateBusinessRequest struct {
 	Description string  `form:"description" json:"description"`
 	Category    string  `form:"category" json:"category" validate:"required"`
 	Address     string  `form:"address" json:"address"`
-	Latitude    float32 `form:"latitude" json:"latitude"`
-	Longitude   float32 `form:"longitude" json:"longitude"`
+	Latitude    float64 `form:"latitude" json:"latitude"`
+	Longitude   float64 `form:"longitude" json:"longitude"`
 	Phone       string  `form:"phone" json:"phone"`
 	Whatsapp    string  `form:"whatsapp" json:"whatsapp"`
 	Instagram   string  `form:"instagram" json:"instagram"`
@@ -19,8 +19,8 @@ type UpdateBusinessRequest struct {
 	Description string  `form:"description" json:"description"`
 	Category    string  `form:"category" json:"category"`
 	Address     string  `form:"address" json:"address"`
-	Latitude    float32 `form:"latitude" json:"latitude"`
-	Longitude   float32 `form:"longitude" json:"longitude"`
+	Latitude    float64 `form:"latitude" json:"latitude"`
+	Longitude   float64 `form:"longitude" json:"longitude"`
 	Phone       string  `form:"phone" json:"phone"`
 	Whatsapp    string  `form:"whatsapp" json:"whatsapp"`
 	Instagram   string  `form:"instagram" json:"instagram"`
